@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 <PropertyGroup>
-    <TargetFramework>net8.0-windows10.0.22621.0</TargetFramework>
+    <TargetFramework>net9.0-windows10.0.22621.0</TargetFramework>
 <RuntimeIdentifiers>win-x64</RuntimeIdentifiers>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -10,10 +10,41 @@
     <LangVersion>preview</LangVersion>
 	<PublishAot>true</PublishAot>
 	<AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+	<PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
+
+<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+  <IsAotCompatible>True</IsAotCompatible>
+  <DefineConstants>$(DefineConstants);RETRO_USE_MOD_LOADER;RETRO_MOD_LOADER_VER_2;RETRO_REV02;RETRO_REV0U</DefineConstants>
+  <DebugType>none</DebugType>
+</PropertyGroup>
+
+<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
+  <IsAotCompatible>True</IsAotCompatible>
+  <DefineConstants>$(DefineConstants);RETRO_USE_MOD_LOADER;RETRO_MOD_LOADER_VER_2;RETRO_REV02;RETRO_REV0U</DefineConstants>
+  <DebugType>none</DebugType>
+</PropertyGroup>
+
+<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+  <IsAotCompatible>True</IsAotCompatible>
+  <DefineConstants>$(DefineConstants);RETRO_USE_MOD_LOADER;RETRO_MOD_LOADER_VER_2;RETRO_REV02;RETRO_REV0U</DefineConstants>
+  <DebugType>none</DebugType>
+</PropertyGroup>
+
+<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
+  <IsAotCompatible>True</IsAotCompatible>
+  <DefineConstants>$(DefineConstants);RETRO_USE_MOD_LOADER;RETRO_MOD_LOADER_VER_2;RETRO_REV02;RETRO_REV0U</DefineConstants>
+  <DebugType>none</DebugType>
+</PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\RSDKv5\RSDKv5.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Using Include="RSDK.EngineAPI">
+      <Static>True</Static>
+    </Using>
   </ItemGroup>
 
 </Project>
