@@ -1,22 +1,9 @@
-﻿using System;
-using System.Drawing;
-using System.Runtime.InteropServices;
-namespace CS_LinkTest;
+﻿namespace CS_LinkTest;
+
 class Program
 {
-
-    // ---------------
-    // RSDKTable types
-    // ---------------
-
-
-
-
     static void Main(string[] args)
     {
-        // do absolutely nothing
-        // (this func prob isnt even needed)
-    }
 
-  
+    }
 }
