@@ -1,8 +1,8 @@
 ﻿#if RETRO_REV02
-namespace RSDK.API;
+namespace RSDK;
 
 public unsafe class RichPresence
 {
-    public static void Set(int id, String* text) => APITable->SetRichPresence(id, text);
+    public static void Set(int id, String text) => APITable.SetRichPresence(id, ref text);
 }
 #endif
