@@ -14,54 +14,36 @@ public unsafe struct String
         size = other.size;
     }
 
+    public void Init(string str, uint length = 0) => RSDKTable.InitString(ref this, str, length);
+    public void Set(string str) => RSDKTable.SetString(ref this, str);
 
-    // TODO: Compare operator
+    // public void Prepend(String* str) { *this = str + *this; }
+    // public void Prepend(string str) { this = new String(str) + this; }
 
+    public void Append(ref String str) => RSDKTable.AppendString(ref this, ref str);
+    public void Append(string str) => RSDKTable.AppendText(ref this, str);
 
-    public void Init(string str, uint length = 0)
-    {
-        fixed (String* s = &this) RSDKTable->InitString(s, str, length);
-    }
+    public static void Copy(String* dst, String* src) => RSDKTable.CopyString(dst, src);
+    public static void Copy(ref String dst, string src) => RSDKTable.SetString(ref dst, src);
+    public static bool32 Compare(String strA, String strB, bool32 exactMatch) { return RSDKTable.CompareStrings(ref strA, ref strB, exactMatch); }
 
-    public void Set(string str)
-    {
-        fixed (String* s = &this) RSDKTable->SetString(s, str);
-    }
+    public void CStr(char* buffer) => RSDKTable.GetCString(buffer, ref this);
 
-    public void Append(String* str) { fixed (String* s = &this) RSDKTable->AppendString(s, str); }
-    public void Append(string str)
-    {
-        var tempStr = new String();
-        tempStr.Set(str);
+    public bool32 Initialized() => chars != null;
+    public bool32 Empty() => length == 0;
 
-        fixed (String* s = &this)
-            RSDKTable->AppendString(s, &tempStr);
-    }
-
-    public static void Copy(String* dst, String* src) => RSDKTable->CopyString(dst, src);
-    public static void Copy(String* dst, string src) => RSDKTable->SetString(dst, src);
-    public static bool32 Compare(String* strA, String* strB, uint exactMatch) { return RSDKTable->CompareStrings(strA, strB, exactMatch); }
-
-    public void CStr(char* buffer) { fixed (String* s = &this) RSDKTable->GetCString(buffer, s); }
-
-    public bool32 Initialized() { return chars != null; }
-    public bool32 Empty() { return length == 0; }
-
-    public void SetSpriteString(SpriteAnimation spriteAni, ushort listID) { fixed (String* s = &this) RSDKTable->SetSpriteString(spriteAni.id, listID, s); }
+    public void SetSpriteString(SpriteAnimation spriteAni, ushort listID) => RSDKTable.SetSpriteString(spriteAni.id, listID, ref this);
     public int GetWidth(SpriteAnimation spriteAni, ushort listID, int spacing)
     {
-        fixed (String* s = &this) return RSDKTable->GetStringWidth(spriteAni.id, listID, s, 0, length, spacing);
+        return RSDKTable.GetStringWidth(spriteAni.id, listID, ref this, 0, length, spacing);
     }
     public int GetWidth(SpriteAnimation spriteAni, ushort listID, int start, int length, int spacing)
     {
-        fixed (String* s = &this) return RSDKTable->GetStringWidth(spriteAni.id, listID, s, start, length, spacing);
+        return RSDKTable.GetStringWidth(spriteAni.id, listID, ref this, start, length, spacing);
     }
 
-    public void LoadStrings(string filepath)
-    {
-        fixed (String* s = &this) RSDKTable->LoadStringList(s, filepath, 16);
-    }
-    public bool32 Split(String* list, int startID, int count) { fixed (String* s = &this) return RSDKTable->SplitStringList(list, s, startID, count); }
+    public void LoadStrings(string filepath) => RSDKTable.LoadStringList(ref this, filepath, 0x10);
+    public bool32 Split(ref String list, int startID, int count) => RSDKTable.SplitStringList(ref list, ref this, startID, count);
 
     public ushort* chars = null;
     public ushort length = 0;
